+++ docs-src/runsite.sh
@@ -1,3 +1,8 @@
 #!/usr/bin/env bash
 
-mkdocs serve
+if ! which hugo &> /dev/null; then
+	echo "Hugo not found. Installing..."
+	brew install hugo
+fi
+
+hugo server -D